move some logic into the HOC
